Fix deadlock during shutdown
Deadlock was caused by one thread calling close on a socket whilst another thread was blocked on a call to recvmsg. In this scenario the recv thread would wake up after the recvmsg had timed out but then attempt another read from the now closed socket. Normally this would result in an error, but if the file descriptor had been reused in the meantime the recvmsg call could end up blocking forever trying to read from the wrong fd.

- Make sure the socket fds used are reset to -1 before closing to prevent other threads attempting to read from a socket that has already been closed.

- Call shutdown on sockets before closing to guarantee other threads reading from the socket are unblocked immediately.

- Check for return of 0 when reading from a socket to handle sockets that have been cleanly closed from another thread.